On SetShaders, set as dirty the SRV, UAV slots used by the program buffers.
On PrepareCompute, bound SRV, UAV for compute shaders.
Note: the same buffer cannot be bound as SRV and UAV at the same time. We could implement an automatic mechanism to unbind shared buffers, but it is complex (some stages can be missing in some commands; compute shaders use different binding functions) and a waste of time (if the render path doesn't change the unbinds to do are always the same). See the old 'shader' branch. Better implement a manual unbind system.